🚚 api: move repository to results context
