+++ api/src/certification/results/infrastructure/repositories/certification-result-repository.js
@@ -1,6 +1,6 @@
-import { knex } from '../../../db/knex-database-connection.js';
-import { ComplementaryCertificationCourseResult } from '../../../src/certification/shared/domain/models/ComplementaryCertificationCourseResult.js';
-import { CertificationResult } from '../../domain/models/CertificationResult.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
+import { CertificationResult } from '../../../../../lib/domain/models/index.js';
+import { ComplementaryCertificationCourseResult } from '../../../shared/domain/models/ComplementaryCertificationCourseResult.js';
 
 const findBySessionId = async function ({ sessionId }) {
   const certificationResultDTOs = await _selectCertificationResults()
